Move UI components into subfolder.
