+++ components/ui/Navbar/Navbar.js
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 import s from './Navbar.module.css';
-import Logo from '../icons/Logo';
-import { useUser } from '../../components/UserContext';
+import Logo from '../../icons/Logo';
+import { useUser } from '../../UserContext';
 
 const Navbar = () => {
   const { user, signOut } = useUser();
